fix: volar lsp broken by rename to 'vue_ls' in mason registry
The update to the mason registry also forced me to update the versions
of lspconfig and mason-lspconfig I had installed, since those were still
configuring the lsp as 'volar', but mason only knew it as 'vue_ls'. This
caused volar to never start whether I had it configured as 'volar' or
'vue_ls', or even had both configured at the same time.